fix: 중복 생성된 teacher entity 통합 및 충돌 해결 - use-dashboard-query.ts import 수정
