+++ src/entities/teacher/infrastructure/teacher.repository.ts
@@ -31,9 +31,9 @@ const transformBasicInfoToFrontend = (
     profilePublicKorean: getProfilePublicKorean(basicInfoDto.isProfilePublic),
   });
 
-/**
+/* ─────────────────────────────────────────────────────
  * [Read] 선생님 기본 정보 조회
- */
+ * ────────────────────────────────────────────────────*/
 const getBasicInfo = async (): Promise<FrontendTeacherBasicInfo> => {
   const response = await api.private.get<CommonResponse<TeacherBasicInfoDTO>>(
     '/teacher/me/basic-info'
@@ -44,9 +44,9 @@ const getBasicInfo = async (): Promise<FrontendTeacherBasicInfo> => {
   return transformBasicInfoToFrontend(basicInfoDto);
 };
 
-/**
+/* ─────────────────────────────────────────────────────
  * [Update] 선생님 기본 정보 변경
- */
+ * ────────────────────────────────────────────────────*/
 const updateBasicInfo = async (
   basicInfo: UpdateTeacherBasicInfoPayload
 ): Promise<void> => {
@@ -54,12 +54,39 @@ const updateBasicInfo = async (
   await api.private.patch('/teacher/me/basic-info', validated);
 };
 
-/**
- * export
- */
+/* ─────────────────────────────────────────────────────
+ * 선생님 통계 조회
+ * ────────────────────────────────────────────────────*/
+const getTeacherReport = async () => {
+  const response = await api.private.get(`/teacher/me/report`);
+  return unwrapEnvelope(response, dto.teacherReport);
+};
+
+/* ─────────────────────────────────────────────────────
+ * 선생님 수업 노트 전체 목록 조회
+ * ────────────────────────────────────────────────────*/
+const getTeacherNoteList = async () => {
+  const response = await api.private.get(`/teacher/me/teaching-notes`);
+  return unwrapEnvelope(response, dto.teacherNoteList);
+};
+
+/* ─────────────────────────────────────────────────────
+ * 선생님 스터디룸 전체 목록 조회
+ * ────────────────────────────────────────────────────*/
+const getTeacherStudyRoomList = async () => {
+  const response = await api.private.get(`/teacher/me/study-rooms`);
+  return unwrapEnvelope(response, dto.teacherStudyRoomList);
+};
+
+/* ─────────────────────────────────────────────────────
+ * 내보내기
+ * ────────────────────────────────────────────────────*/
 export const repository = {
   basicInfo: {
     getBasicInfo,
     updateBasicInfo,
   },
+  getTeacherNoteList,
+  getTeacherStudyRoomList,
+  getTeacherReport,
 };
