--- src/entities/teacher/infrasturcture/index.ts
@@ -1,3 +1,0 @@
-export * from './teacher.repository';
-export * from './teacher.keys';
-export * from './teacher.dto';
