--- src/entities/teacher/infrasturcture/teacher.dto.ts
@@ -1,49 +1,0 @@
-import { z } from 'zod';
-
-/* ─────────────────────────────────────────────────────
- * 선생님 통계 조회 DTO
- * ──────────────────────────────────────────────────── */
-const TeacherReportDtoSchema = z.object({
-  studyRoomCount: z.number(),
-  teachingNoteCount: z.number(),
-  studentCount: z.number(),
-  qnaCount: z.number(),
-  reviewCount: z.number(),
-});
-
-/* ─────────────────────────────────────────────────────
- * 선생님 수업 노트 전체 목록 조회 DTO
- * ────────────────────────────────────────────────────*/
-const TeacherNoteListItemDtoSchema = z.object({
-  id: z.number(),
-  title: z.string(),
-  studyRoomId: z.number(),
-  studyRoomName: z.string(),
-  qnaCount: z.number(),
-  viewCount: z.number(),
-  modDate: z.string(),
-  representative: z.boolean(),
-});
-
-const TeacherNoteListDtoSchema = z.array(TeacherNoteListItemDtoSchema);
-
-/* ─────────────────────────────────────────────────────
- * 선생님 스터디룸 전체 목록 조회 응답 DTO
- * ────────────────────────────────────────────────────*/
-const TeacherStudyRoomListItemDtoSchema = z.object({
-  id: z.number(),
-  name: z.string(),
-  teachingNoteCount: z.number(),
-  studentCount: z.number(),
-  qnaCount: z.number(),
-});
-
-const TeacherStudyRoomListDtoSchema = z.array(
-  TeacherStudyRoomListItemDtoSchema
-);
-
-export const dto = {
-  teacherReport: TeacherReportDtoSchema,
-  teacherNoteList: TeacherNoteListDtoSchema,
-  teacherStudyRoomList: TeacherStudyRoomListDtoSchema,
-};
