--- src/entities/teacher/infrasturcture/teacher.keys.ts
@@ -1,6 +1,0 @@
-export const teacherKeys = {
-  all: ['teacher'] as const,
-  report: () => [...teacherKeys.all, 'report'] as const,
-  noteList: () => [...teacherKeys.all, 'noteList'] as const,
-  studyRoomList: () => [...teacherKeys.all, 'studyRoomList'] as const,
-};
